Initial Commit: SYSC4504 Assignment 1 - Creating SYSCX Website Pages
This commit includes the foundational work for SYSC4504's first assignment. It involves the creation of three main web pages intended to simulate a social media platform similar to Twitter for SYSC students. The pages developed are:

index.html - The home page setup.
profile.html - User profile configuration page with forms and a drop-down menu for program selection.
register.html - New user registration form.

All pages are styled using CSS and include forms that simulate interaction with backend scripts via predefined .php files (not functional in this stage). The project was built following the assignment's specifications to use HTML and CSS basics, with a focus on form and table usage as per the course teachings.